% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockMatrix.R
\name{blockMatrix}
\alias{blockMatrix}
\title{blockMatrix}
\usage{
blockMatrix(nrow, ncol, nRowBlock, nColBlock, signalBlock = NULL)
}
\arguments{
\item{nrow}{number of rows}

\item{ncol}{number of columns}

\item{nRowBlock}{number of blocks in rows and cols}

\item{nColBlock}{number of blocks in rows and cols}

\item{signalBlock}{matrix \code{nRowBlock} x \code{nColBlock} of parameters used to generate
the matrix entries within each block}
}
\value{
list containing the following
\item{mat}{block matrix}
\item{signalBlock}{see input parameters}
\item{idRows}{vector indicating the division of rows into blocks}
\item{idRows}{vector indicating the division of columns into blocks}
}
\description{
Generation by block with gamma distribution,
each block has specific parameter values
}
\details{
The matrix is constructed by blocks in rows and columns,
the entries of each block are constant and specified in input
}
\author{
Ghislain Durif, \email{gd.dev@libertymail.net}
}
