% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGeneration.R
\name{dataGeneration}
\alias{dataGeneration}
\title{dataGeneration}
\usage{
dataGeneration(n, p, K, alpha1, alpha2, beta1, beta2, ZI = FALSE,
  prob0 = NULL)
}
\arguments{
\item{n}{number of observations (nb of rows)}

\item{p}{number of variables (nb of columns)}

\item{K}{number of latent factors (dimension of latent subspace)}

\item{alpha1}{matrix n x K, first parameters for the prior distribution on U}

\item{beta1}{matrix n x K, first parameters for the prior distribution on V}

\item{beta2}{matrix n x K, second parameters for the prior distribution on U}

\item{ZI}{boolean, indicating if the data are zero-inflated (default is FALSE)}

\item{prob0}{vector of Bernoulli probability for zero-inflation (default is NULL), length p, one probability per variable}

\item{alpĥa2}{matrix n x K, second parameters for the prior distribution on U}
}
\value{
list containing the following
\item{X}{data matrix n x p of counts}
\item{U}{matrix n x K of factor coordinates (in observation space)}
\item{V}{matrix p x K of factor loadings (in variable space)}
\item{n}{number of observations (or rows in X)}
\item{p}{number of variables (or columns in X)}
\item{K}{number of latent factors (dimension of latent subspace)}
\item{alpha1}{matrix n x K, first parameters for the prior distribution on U}
\item{alpĥa2}{matrix n x K, second parameters for the prior distribution on U}
\item{beta1}{matrix n x K, first parameters for the prior distribution on V}
\item{beta2}{matrix n x K, second parameters for the prior distribution on U}
\item{ZI}{boolean, indicating if the data are zero-inflated (default is FALSE)}
\item{prob0}{vector of Bernoulli probability for zero-inflation}
}
\description{
Simulates a data matrix X according to the (possibly
zero-inflated) Gamma Poisson Factor Model
}
\details{
## generative process (without zero-inflation)
  X | U,V ~ Poisson(U t(V))
  U ~ Gamma(alpha)
  V ~ Gamma(beta)

## Generative process (with zero-inflation)
 X_{ij} = sum_k Z_{ijk}
 X_{ij} | U,V ~ Y_{ij} * Poisson(U_i t(V_j))
 Z_{ijk} | U,V ~ Y_{ij} * Poisson(U_{ik} V_{jk})
 Y_{ij} ~ Bernoulli(p_j)
 U ~ Gamma(alpha)
 V ~ Gamma(beta)
}
\author{
Ghislain Durif, \email{ghislain.durif@univ-lyon1.fr}
}

