% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{deviance}
\alias{deviance}
\title{deviance}
\usage{
deviance(model, X)
}
\arguments{
\item{model}{a Gamma-Poisson factor model output by the
function \code{\link{pCMF}}}
}
\value{
the matrix U of individuals coordinates in the lower
dimensional sub-space
}
\description{
Get the associated to a Gamma-Poisson Factor model
}
\details{

}
\seealso{
\code{\link{expDev}}
}
\author{
Ghislain Durif, \email{gd.dev@libertymail.net}
}
