% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{expDev}
\alias{expDev}
\title{expDev}
\usage{
expDev(model, X)
}
\arguments{
\item{model}{a Gamma-Poisson factor model output by the
function \code{\link{pCMF}}}
}
\value{
the matrix U of individuals coordinates in the lower
dimensional sub-space
}
\description{
Get the percentage of deviance explained by a Gamma-Poisson Factor model
}
\details{
see pCMF function output
}
\seealso{
\code{\link{pCMF}}
}
\author{
Ghislain Durif, \email{gd.dev@libertymail.net}
}
