% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gamPoisFactorPen_wrapper}
\alias{gamPoisFactorPen_wrapper}
\title{gamPoisFactorPen_wrapper}
\usage{
gamPoisFactorPen_wrapper(Xin, K, ZI, phi01in, phi02in, theta01in, theta02in,
  alpha1in, alpha2in, beta1in, beta2in, lambda_in, mu_in, iterMax, epsilon,
  order, stabRange, verbose)
}
\arguments{
\item{K}{number of factors}

\item{X}{matrix n x p of counts}

\item{phi01}{n x K, initial values of first parameter of Gamma distribution on U}

\item{phi02}{n x K, initial values of second parameter of Gamma distribution on U}

\item{theta01}{n x K, initial values of first parameter of Gamma distribution on V}

\item{theta02}{n x K, initial values of second parameter of Gamma distribution on V}

\item{alpha1}{n x K, initial values of first parameter of Gamma prior on U}

\item{alpha2}{n x K, initial values of second parameter of Gamma prior on U}

\item{beta1}{n x K, initial values of first parameter of Gamma prior on V}

\item{beta2}{n x K, initial values of second parameter of Gamma prior on V}
}
\value{
return
\item{Y}{Y}
}
\description{
Description
}
\details{
Wrapper for Cpp function
}
\author{
Ghislain Durif, \email{ghislain.durif@univ-lyon1.fr}
}
\seealso{
aaa
}
\keyword{internal}

