% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getLogV}
\alias{getLogV}
\title{getLogV}
\usage{
getLogV(model)
}
\arguments{
\item{model}{a Gamma-Poisson factor model output by the
function \code{\link{pCMF}}}
}
\value{
the matrix V of variables \(features\) contributions to the lower
dimensional sub-space in the geometry related to the Gamma distribution
in the exponential family (log)
}
\description{
Getter for the log of matrix V in the Gamma-Poisson Factor model
}
\details{
see pCMF function output
}
\seealso{
\code{\link{pCMF}}
}
\author{
Ghislain Durif, \email{gd.dev@libertymail.net}
}
