% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getU}
\alias{getU}
\title{getU}
\usage{
getU(model)
}
\arguments{
\item{model}{a Gamma-Poisson factor model output by the
function \code{\link{pCMF}}}
}
\value{
the matrix U of individuals coordinates in the lower
dimensional sub-space
}
\description{
Getter for the matrix U in the Gamma-Poisson Factor model
}
\details{
see pCMF function output
}
\seealso{
\code{\link{pCMF}}
}
\author{
Ghislain Durif, \email{gd.dev@libertymail.net}
}
