% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{graphU}
\alias{graphU}
\title{graphU}
\usage{
graphU(model, axes = c(1, 2), labels = NULL,
  log_representation = TRUE, edit_theme = TRUE, graph = TRUE)
}
\arguments{
\item{model}{model a Gamma-Poisson factor model output by the
function \code{\link{pCMF}} or a Poisson NMF model output by the
function \code{\link{run_poisson_nmf}}}

\item{axes}{a vector of 2 indexes corresponding to the 2 directions
to represent, Default value is \code{c(1,2)}.}

\item{labels}{a vector of indidividual/feature labels. Default value
is NULL, in this case individuals/features are not labelled.}

\item{log_representation}{boolean, indicating if the representation is in
the natural geometry associated to the Gamma distribution (log) or in the
standard Euclidean space. Default value is TRUE.}

\item{edit_theme}{boolean, indicating if the ggplot2 standard theme
should be edited or not. Default value is TRUE.}

\item{graph}{boolean, indicating if the graph should be drawn or not.
Default value is TRUE.}
}
\value{
the ggplot2 graph
}
\description{
Graphical representation of the individuals coordinates (matrix U)
in the pCMF/NMF framework (\eqn{X ~ UV^t})
}
\details{
We consider a count data matrix \eqn{X} with
\eqn{n} observations/individuals in rows and
\eqn{p} recorded variables in columns.

In the matrix factorization framework (Poisson NMF or Gamma-Poisson
factor model), the data matrix \eqn{X_{n\times p}} is approximated by the
product \eqn{UV^t} where \eqn{U_{n \times K}} and \eqn{V_{p\times K}}
are respectively the observation coordinates and variable loadings in
the latent space of lower dimension \eqn{K}.

This function can be used to draw the scatterplot of individual coordinates
along two of the latent directions, i.e. two columns from the
matrix \code{U}.

In the case of the Gamma-Poisson factor model, the geometry related to the
Gamma distribution in the exponential family is a log-representation,
thus, if requested, the matrix \code{log U} can be returned.

The graphical representation is based on ggplot2.
}
\examples{
\dontrun{
## generate data
n <- 100
p <- 200
K <- 10
factorU <- generate_factor_matrix(n, K, ngroup=3, average_signal=60,
                                  group_separation=0.8,
                                  distribution="gamma",
                                  shuffle_feature=TRUE)
factorV <- generate_factor_matrix(p, K, ngroup=2, average_signal=60,
                                  group_separation=0.8,
                                  distribution="gamma",
                                  shuffle_feature=TRUE)
U <- factorU$factor_matrix
V <- factorV$factor_matrix
count_data <- generate_count_matrix(n, p, K, U, V)
X <- count_data$X
## or use your own data as a count matrix
## of dimension cells x genes (individuals x features)
## run pCMF algorithm
res <- pCMF(X, K=2, verbose=FALSE)
## graphical representation
graphU(res, axes=c(1,2), labels=factorU$feature_label) # individual representation
graphV(res, axes=c(1,2), labels=factorV$feature_label) # variable representation
}

}
\seealso{
\code{\link{pCMF}}, \code{\link{run_poisson_nmf}}
}
\author{
Ghislain Durif, \email{gd.dev@libertymail.net}
}
