% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_heatmap.R
\name{matrix_heatmap}
\alias{matrix_heatmap}
\title{matrix_heatmap}
\usage{
matrix_heatmap(mat, ...)
}
\arguments{
\item{mat}{the matrix to plot}

\item{...}{any parameter that can be fed to the function
\code{\link[fields]{image.plot}}}
}
\description{
Wrapper for the function \code{\link[fields]{image.plot}} from the
package \code{fields}
}
\details{
Plot the entries of the input matrix as a heatmap
}
\examples{
\dontrun{
n <- 100
K <- 50
mat <- generate_factor_matrix(n, K, ngroup=3, average_signal=0.5*K,
                              group_separation=0.5,
                              distribution="uniform",
                              shuffle_feature=FALSE)
matrix_heatmap(mat$factor_matrix)
}

}
\author{
Ghislain Durif, \email{gd.dev@libertymail.net}
}
