% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{matrix_plot}
\alias{matrix_plot}
\title{matrix_plot}
\usage{
matrix_plot(mat, axes = c(1:2), labels = NULL, edit_theme = TRUE,
  graph = TRUE)
}
\arguments{
\item{mat}{a real-valued matrix of dimension \code{nrow x K} representing
the coordinates of individuals/features (rows) along \code{K} latent
directions (columns).}

\item{axes}{a vector of 2 indexes corresponding to the 2 directions
to represent, Default value is \code{c(1,2)}.}

\item{labels}{a vector of indidividual/feature labels. Default value
is NULL, in this case individuals/features are not labelled.}

\item{edit_theme}{boolean, indicating if the ggplot2 standard theme
should be edited or not. Default value is TRUE.}

\item{graph}{boolean, indicating if the graph should be drawn or not.
Default value is TRUE.}
}
\value{
the ggplot2 graph
}
\description{
Graphical representation of a bidimensional scatterplot based on columns of
the input matrix \code{mat}.
}
\details{
This function is a wrapper to construct a ggplot2 graph representing
coordinates of individuals or features embedded in a \code{K} dimensional
latent space. Each row of the input matrix \code{mat} represents an
individual or a feature, and each column represents a latent dimension.

This function plots the point coordinates (rows) from the matrix \code{mat}
according to 2 chosen directions (columns). Default behavior is to use the
first two direction (columns).

This function is internal and is used by the function \code{\link{graphU}}
and \code{\link{graphV}}.
}
\seealso{
\code{\link{graphU}} \code{\link{graphV}}
}
\author{
Ghislain Durif, \email{gd.dev@libertymail.net}
}
\keyword{internal}
