% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixHeatmap.R
\name{matrixHeatmap}
\alias{matrixHeatmap}
\title{matrixHeatmap}
\usage{
matrixHeatmap(mat, ...)
}
\arguments{
\item{mat}{the matrix to plot}
}
\description{
plot the entries of a matrix as a heatmap
}
\details{
hello world
}
\author{
Ghislain Durif, \email{ghislain.durif@univ-lyon1.fr}
}

