% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{matrixPlot}
\alias{matrixPlot}
\title{matrixPlot}
\usage{
matrixPlot(mat, axes = c(1:2), labels = NULL, log_representation = TRUE,
  edit_theme = TRUE, graph = TRUE)
}
\arguments{
\item{axes}{a vector of 2 indexes corresponding to the 2 axes to represent}

\item{labels}{a vector of indidividual labels}

\item{log_representation}{boolean, indicating if the representation is in
the natural geometry associated to the Gamma distribution (log) or in the
Euclidean space, default is TRUE}

\item{edit_theme}{boolean, indicating if the ggplot2 standard should be
edited or not}

\item{graph}{boolean, indicating if the graph should be drawn or not}

\item{model}{a Gamma-Poisson factor model output by the
function \code{\link{pCMF}}}
}
\value{
the matrix U of individuals coordinates in the lower
dimensional sub-space
}
\description{
Graphical representation of the variables contributions (matrix V)
in the Gamma-Poisson Factor model
}
\details{
Coordinates of the variables in the Gamma-Poisson factor model in a 2-D
space, along two of the axes in the dimensional sub-space
in the Euclidean geometry, corresponding to the
matrix V, or in the geometry related to the Gamma distribution in the
exponential family (log), corresponding to matrix logV

Graphical representation based on ggplot2

see pCMF function output
}
\seealso{
\code{\link{pCMF}}
}
\author{
Ghislain Durif, \email{gd.dev@libertymail.net}
}
\keyword{internal}
