% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCMF.R
\docType{package}
\name{pCMF.package}
\alias{pCMF.package}
\alias{pCMF.package-package}
\title{pCMF.package}
\description{
The pCMF package implements algorithms for factorization of count matrices
based on probabilistic models, especially Gamma-Poisson factor models (with
variants including zero-inflation and sparsity). It was specifically
designed to analyse single-cell gene expression data (scRNA-seq).
}
\details{
R package pCMF

See \url{https://gitlab.inria.fr/gdurif/pCMF}


More details can be found in Durif et al. (2017). Matrix factorization is an
unsupervised dimension reduction approach, suitable for data visualization
or as a preliminary step for clustering.
}
\references{
Durif, G., Modolo, L., Mold, J.E., Lambert-Lacroix, S., Picard, F., 2017.
Probabilistic Count Matrix Factorization for Single Cell Expression Data
Analysis. arXiv:1710.11028 [stat].
}
\author{
Ghislain Durif \email{gd.dev@libertymail.net}

Maintainer: Ghislain Durif \email{gd.dev@libertymail.net}
}
