% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{poisLogLike}
\alias{poisLogLike}
\title{poisLogLike}
\usage{
poisLogLike(X, lambda)
}
\arguments{
\item{X}{matrix n x p of count data}

\item{lambda}{matrix n x p of Poisson rate}
}
\value{
value of the Poisson log-likelihood
}
\description{
Poisson log-likelihood
}
\details{

}
\author{
Ghislain Durif, \email{gd.dev@libertymail.net}
}
\keyword{internal}
