% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch-routine-registration.R
\name{run_testthat_tests}
\alias{run_testthat_tests}
\title{run_testthat_tests}
\description{
This dummy function definition is included with the package to ensure that
`tools::package_native_routine_registration_skeleton()` generates the
required registration info for the 'run_testthat_tests' symbol.
}
\details{
See \url{https://github.com/r-lib/testthat} or
\url{https://CRAN.R-project.org/package=testthat}
}
\author{
Hadley Wickham; RStudio
}
\keyword{internal}
