% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wrapper_varEM_ZI_GaP}
\alias{wrapper_varEM_ZI_GaP}
\title{wrapper_varEM_ZI_GaP}
\usage{
wrapper_varEM_ZI_GaP(Xin, K, ZI, phi01in, phi02in, theta01in, theta02in,
  alpha1in, alpha2in, beta1in, beta2in, iterMax, iterMin, epsilon, order,
  stabRange, verbose, ncores, nbInit, iterMaxInit, noise, seed)
}
\arguments{
\item{K}{dimension of the latent subspace}

\item{ZI}{boolean, indicating if the model is zero-inflated}

\item{iterMax}{maximum number of iterations}

\item{iterMin}{minimum number of iterations}

\item{epsilon}{precision for comparison when assessing convergence}

\item{order}{derivative order on normalized gap to assess convergence
(0 is the current value, 1 the first order empirical derivative, 2 the second order empirical derivative)}

\item{stabRange}{range of stability (number of iterations where parameter values are stable to confirm convergence)}

\item{verbose}{boolean indicating verbosity in the output}

\item{ncores}{number of cores to use for multi-threading}

\item{nbInit}{number of initialization to try}

\item{iterMaxInit}{number of iteration to for each initialization in multi-initialization case (nbInit > 1)}

\item{noise}{nosie level (between 0 and 1) for multi-init case}

\item{seed}{for RNG (-1 means no seed)}

\item{X}{n x p, count data matrix}

\item{phi1}{n x K, initial values of first parameter of Gamma distribution on U}

\item{phi2}{n x K, initial values of second parameter of Gamma distribution on U}

\item{theta1}{n x K, initial values of first parameter of Gamma distribution on V}

\item{theta2}{n x K, initial values of second parameter of Gamma distribution on V}

\item{alpha1}{n x K, initial values of first parameter of Gamma prior on U}

\item{alpha2}{n x K, initial values of second parameter of Gamma prior on U}

\item{beta1}{n x K, initial values of first parameter of Gamma prior on V}

\item{beta2}{n x K, initial values of second parameter of Gamma prior on V}
}
\value{
return
\item{Y}{Y}
}
\description{
Description
}
\details{
Wrapper for Cpp function
}
\seealso{
aaa
}
\author{
Ghislain Durif, \email{gd.dev@libertymail.net}
}
\keyword{internal}
