% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_factor_matrix}
\alias{generate_factor_matrix}
\title{generate_factor_matrix}
\usage{
generate_factor_matrix(nfeature, K, ngroup = 1, average_signal = 1,
  disp = NULL, group_separation = 0.5, distribution = "uniform",
  shuffle_feature = TRUE, prop_noise_feature = 0, noise_level = NULL,
  seed = NULL, tag = NULL)
}
\arguments{
\item{nfeature}{integer, number of features (individuals or variables),
i.e. number of rows in the generated matrix.}

\item{K}{integer, dimension of the latent sub-space, i.e. number of columns
in the generated matrix.}

\item{ngroup}{integer, number of groups structuring the features
(individuals or variables). Such groups induce a structure of dependence
between the features. Default value is 1 and the features are not
structured into dependent groups. Condition: \code{ngroup} <= \code{K}.}

\item{average_signal}{real value or vector of real value of size \code{ngroup},
average signal level in the structured block of the generated matrix. If a
single value, all groups share the same \code{average_signal}. If a
vector, each group has the corresponding \code{average_signal}.}

\item{disp}{real value or vectorof real value of size \code{ngroup}.
If distribution is used, \code{disp} defines the dispersion parameter
for each/all group(s) (if resp vector/single value) for the gamma distribution.
See details. Default value is NULL, and \code{disp} is set to 1.}

\item{group_separation}{real value in \[0,1\] indicating the level of
separation between the groups, 0 = no separation, 1 = high separation.
Default value is 0.5.}

\item{distribution}{character string, distribution used to generate
the matrix entries. Default is "uniform", other possible values are
"gamma" and "exponential". See details.}

\item{shuffle_feature}{boolean, should the features (individuals
or variables) be shuffled in the output.}

\item{prop_noise_feature}{real value in \[0,1\], proportion of features
that does not contributes to the group structure, i.e. constitute an
additional group of noise. Default is 0.}

\item{noise_level}{real value, average level of the noisy feature. Default
is null and the level of noisy feature is set to be
(1 - \code{group_separation}) * \code{average_signal} / \code{disp}.}

\item{seed}{integer, seed value for random number generator. Default value
is NULL and no the seed is not specifically set.}

\item{tag}{string, tag to name feature (i.e. rows in the factor matrix),
with "tagXXX" where "XXX" is the row number. Default value is NULL and
feature are labelled "1", "2", etc.}
}
\value{
list containing the following
\item{factor_matrix}{the factor matrix of dimension \code{nfeature x K}.}
\item{feature_order}{vector of index giving the shuffling order used to
generate the data}
\item{feature_label}{vector of labels}
}
\description{
Generation of a factor matrix in a latent sub-space of dimension \code{K}
}
\details{
This function can be used to generate factor matrix in
a matrix factorization framework \eqn{X ~ UV^t}:
    - U for individual representation in the latent sub-space
    - V for variable contribution to the latent sub-space

More details regarding data generation can be found in Durif et al. (2017,
Appendix).

If the uniform distribution is used, entries in \code{U} and \code{V} are
simulated from the distribution Uniform(0, 2*\code{average_signal}/\code{K}).

If the gamma distribution is used, entries in \code{U} and \code{V} are
simulated from the distribution Gamma(\code{average_signal}/\code{K}, 1).
If the parameter \code{disp} is given, then the distribution
Gamma(\code{average_signal}/\code{K}, \code{disp}) is used.

If the exponential distribution is used, entries in \code{U} and \code{V} are
simulated from the distribution Gamma(1, \code{K}/\code{average_signal})
which corresponds to the distribution
Exponential(\code{K}/\code{average_signal}).

The noise variables are generated with the same framework, using
the parameter \code{noise_level} instead of \code{average_signal}.
}
\examples{
\dontrun{
## generate data
n <- 100
K <- 5
factorU <- generate_factor_matrix(n, K, ngroup=3, average_signal=0.5*K,
                                  group_separation=0.5,
                                  distribution="uniform",
                                  shuffle_feature=TRUE)
}

}
\references{
Durif, G., Modolo, L., Mold, J.E., Lambert-Lacroix, S., Picard, F., 2017.
Probabilistic Count Matrix Factorization for Single Cell Expression Data
Analysis. arXiv:1710.11028 [stat].
}
\author{
Ghislain Durif, \email{gd.dev@libertymail.net}
}
