% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFactor.R
\name{matrixFactor}
\alias{matrixFactor}
\title{matrixFactor}
\usage{
matrixFactor(X, K, phi01, phi02, theta01, theta02, alpha1, alpha2, beta1, beta2,
  lambda = NULL, mu = NULL, iterMax = 200, iterMin = 1,
  epsilon = 1e-05, order = 0, stabRange = 5, verbose = TRUE,
  sparse = FALSE, ZI = FALSE, algo = "EM")
}
\arguments{
\item{X}{matrix n x p of counts}

\item{K}{number of factors}

\item{phi01}{n x K, initial values of first parameter of Gamma distribution on U}

\item{phi02}{n x K, initial values of second parameter of Gamma distribution on U}

\item{theta01}{n x K, initial values of first parameter of Gamma distribution on V}

\item{theta02}{n x K, initial values of second parameter of Gamma distribution on V}

\item{alpha1}{n x K, initial values of first parameter of Gamma prior on U}

\item{alpha2}{n x K, initial values of second parameter of Gamma prior on U}

\item{beta1}{n x K, initial values of first parameter of Gamma prior on V}

\item{beta2}{n x K, initial values of second parameter of Gamma prior on V}
}
\value{
return
\item{Y}{Y}
}
\description{
R wrapper for Gamma-Poisson Factor model
}
\details{
Factorization of count matrix
}
\seealso{
aaa
}
\author{
Ghislain Durif, \email{gd.dev@libertymail.net}
}
