% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCMF.R
\name{pCMF}
\alias{pCMF}
\title{pCMF}
\usage{
pCMF(X, K, alpha1 = NULL, alpha2 = NULL, beta1 = NULL, beta2 = NULL,
  iterMax = 200, iterMin = 100, epsilon = 1e-05, verbose = TRUE,
  sparse = FALSE, ZI = FALSE, ncores = 1)
}
\arguments{
\item{X}{matrix n x p of counts}

\item{K}{number of factors}

\item{alpha1}{n x K, initial values of first parameter of Gamma prior on U}

\item{alpha2}{n x K, initial values of second parameter of Gamma prior on U}

\item{beta1}{n x K, initial values of first parameter of Gamma prior on V}

\item{beta2}{n x K, initial values of second parameter of Gamma prior on V}
}
\value{
return
\item{Y}{Y}
}
\description{
R wrapper for Gamma-Poisson Factor model
}
\details{
Factorization of count matrix
}
\seealso{
aaa
}
\author{
Ghislain Durif, \email{gd.dev@libertymail.net}
}
